feat: refactor options file location to table creator (#302)
