--- src/components/TableCreator/hooks/useTableOptions/types.ts
@@ -1,3 +1,0 @@
-import { RowData, TableOptions } from "@tanstack/react-table";
-
-export type UseTableOptions<T extends RowData> = Partial<TableOptions<T>>;
